### 2021-03-25 #3135 @Kevin-Mattheus-Moerman
@tresoldi thanks for this submission. I've flagged this submission for additional scope review by our editorial board. This is default for submissions of this size (see our ["substantial scholarly effort" requirements](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)) and should take about a week to complete. Thanks for your patience.
