### 2021-04-08 #3135 @majensen
@DavidNickle - would you have some time to review [this interesting work on phylogenetic simulation](#3135) for [JOSS](https://joss.theoj.org)?
