### 2021-04-08 #3135 @DavidNickle
Hi Mark,

I would be happy to but the current link does not work.

Thanks,

David

> On Apr 8, 2021, at 11:23 AM, Mark Jensen ***@***.***> wrote:
>
>
> @DavidNickle <https://github.com/DavidNickle> - would you have some time to review this interesting work on phylogenetic simulation <x-msg://64/#3135> for JOSS <https://joss.theoj.org/>?
>
