### 2021-04-11 #3135 @majensen
@rvosa would you be up for reviewing [this phylogenetic simulation library](#3135) for [JOSS](https://joss.theoj.org)? A guide for JOSS reviews are [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

It would be great to have you weigh in! Hope things are great for you --@majensen
