### 2021-04-15 #3173 @whedon
[REVIEW]: Ngesh: a Python library for synthetic phylogenetic data
**Submitting author:** @tresoldi (<a href="http://orcid.org/0000-0002-2863-1467">Tiago Tresoldi</a>)
**Repository:** <a href="https://github.com/tresoldi/ngesh" target ="_blank">https://github.com/tresoldi/ngesh</a>
**Version:** v0.5
