### 2021-04-15 #3173 @DavidNickle
**Conflict of interest:** I confirm that I have read the JOSS conflict of interest policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
**Code of Conduct**

**Repository:** The source code for this software is available at the repository url **YES**

**License:** The repository contain a plain-text LICENSE file with the contents of an OSI approved software license. **YES**

**Contribution and authorship:** The submitting author made major contributions to the software and the full list of paper authors seem appropriate and complete. **YES**

**Functionality:** The software is functional and important adding very important knowledge to the public domain. **YES**

**Installation:** The installation procedure follows the documentation clearly.  I performed many of the examples in the How To Use section of the manuscript.  All worked as advertised

**Functionality:** I confirmed the functional claims of the software as well as some of my own.

**Performance:** It is pretty quick!

**A statement of need:** The authors have a clear audience in they have mind and the paper and the software should be readably available to that audience. It is potentially most useful for developing nulls in the evolution of languages

**Installation instructions:** I installed through PIP without any issues

**Example usage:** I went through the How To Use section. Everything worked.

**Automated tests:** YES — tests are available

**Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. **YES**

**Software paper:**

**Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided? **YES**

**A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is? **YES**

**State of the field:** Do the authors describe how this software compares to other commonly-used packages? YES
Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)? YES
References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax? **YES**

@majensen I think this paper should be accepted. It is a nice piece of software that can be useful for those that are interested in the evolutionary of languages. It even allows for recombination (words passing from one language into another — can you use the word **pajama** or **pyjama**).
