### 2021-05-09 #3173 @rvosa
- [x] I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
- [x] I confirm that I read and will adhere to the JOSS code of conduct.
- [x] Yes: the source code for this software is available at the repository url?
- [x] Yes: the repository contains a plain-text (MIT) LICENSE file with the contents of an OSI approved software license?
- [x] Yes: the submitting author (@tresoldi) has made all but one of the commits. Their sole authorship therefore seems appropriate.
- [x] Yes: this submission meets the scope eligibility described in the JOSS guidelines
- [x] Yes: the installation proceeds more or less as outlined in the documentation, though I needed to use `pip3` instead. I blame my messy Python configuration though.
- [x] Yes: the functional claims of the software have been confirmed. I type `ngesh` and I get a b/d tree.
- [x] There were no explicit performance claims but also no issues.
- [ ] Maybe: there is no explicit statement of need but the Background section in the PDF makes the case reasonably well. I do note that the presented functionality is available in many other toolkits as well and as such does not constitute a major contribution.
- [x] Yes: there is a clearly-stated list of dependencies in the form of a `requirements.txt`
- [x] Yes: the authors include examples of how to use the software
- [x] Yes: the core functionality of the software is documented to a satisfactory level (e.g., API method documentation)?
- [x] Yes: there are automated tests so that the functionality of the software can be verified?
- [x] Yes: there are clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
- [x] Yes: a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience has been provided
- [ ] **No**: the paper has no section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is
- [ ] **No**: the authors do not describe sufficiently how this software compares to other commonly-used packages. Simulating trees and characters is available in numerous other tools, e.g. DendroPy, Mesquite, Bio::Phylo and others
- [x] Yes: the paper is well written (i.e., it does not require editing for structure, language, or writing quality)
- [ ] Maybe: the list of references is complete in that it has the citations in the manuscript, but those are incomplete, as noted above.
