### 2021-05-11 #3173 @majensen
@tresoldi - @rvosa would like to see some text comparing your package to other available tools, and to highlight how your software meets an existing need. JOSS does requirement this so-called "Statement of Need". Can you work to add this? thanks
