### 2021-05-26 #3173 @tresoldi
@majensen I have finished reworking the draft. The last paragraph of the "Background" was reworked into a "Statement of need", with no real changes in what is said. I also removed some filler words from the rest of the draft to make sure it stays within 3 pages when compiled, but contents are otherwise unchanged.
