### 2021-06-14 #3173 @majensen
Sorry for the continued delay @tresoldi -- finally able to have a look.

What I appreciate about the paper is the idea of "technology transfer" of phylogenetic methods from biology to linguistics. (Although, I think biology imported the basic ideas from comparative linguistics a long long time ago!)

I have to agree with @rvosa, however, that there are many and extensive tools that have been developed for simulating trees in the evolutionary biology world - see https://evolution.genetics.washington.edu/phylip/software.html#Simulation for an old list. Now, the fact that there are about 30 listed there suggests that people need certain things for their applications that aren't quite there in other programs, so I am not inclined to dismiss your submission just because others have programs (in other languages or with more specialized algorithms) that perform similar functions.

However - I would like a little more discussion comparing your application with a few of these common evolutionary biology tools. Can I ask you to review the ones suggested by @rvosa (DendroPy, Mesquite, and Bio::Phylo) and compare your application with these. I'm not too concerned with extending the length of the paper. I would also suggest briefly outlining a real workflow in the linguistic space, one that you have used ngesh for, perhaps. This would be helpful to readers in that discipline and I think it would help argue for the usefulness of your contribution.

Sorry to make more work after such a long delay, but I think you'll have a better paper at the end. Plus, we have @rvosa to convince!
Thanks for your consideration -
