+++ manuscript/paper.bib
@@ -1,3 +1,20 @@
+@misc{sukumaran2021,
+	author = {Sukumaran, J. and Mark T. Holder},
+	title = {The DendroPy Phylogenetic Computing Library Documentation. Retrieved Apr 08, 2021},
+	url = {http://dendropy.org/}
+}
+
+@article{cldf,
+    title={Cross-Linguistic Data Formats, advancing data sharing and re-use in comparative linguistics},
+    author={Forkel, Robert and List, Johann-Mattis and Greenhill, Simon J and Rzymski, Christoph and Bank, Sebastian and Cysouw, Michael and Hammarstr{\"o}m, Harald and Haspelmath, Martin and Kaiping, Gereon A and Gray, Russell D},
+    journal={Scientific Data},
+    volume={5},
+    number={1},
+    pages={1--10},
+    year={2018},
+    publisher={Nature Publishing Group}
+}
+
 @book{Bailey:1990,
   title={The elements of stochastic processes with applications to the natural sciences},
   author={Bailey, Norman T.J.},
