+++ manuscript/paper.bib
@@ -1,6 +1,7 @@
-@misc{sukumaran2021,
+@misc{Sukumaran2021,
 	author = {Sukumaran, J. and Mark T. Holder},
 	title = {The DendroPy Phylogenetic Computing Library Documentation. Retrieved Apr 08, 2021},
+	year = {2021},
 	url = {http://dendropy.org/}
 }
 
