### 2021-10-03 #3173 @majensen
Hi @tresoldi - it is looking good to me. I would still like to see a few sentences describing a typical analysis in historical linguistics and where ngesh fits in to that analysis.

Can you update your paper.bib with the DOIs found by Whedon as MISSING DOIs [above](https://github.com/openjournals/joss-reviews/issues/3173#issuecomment-933022171).

Also, can you update the reference to Harmon's book with more precise details as follows:
- title: Phylogenetic Comparative Methods: Learning from Trees
- year: 2018
- isbn: 978-1719584463
- publisher: CreateSpace Independent Publishing (not University of Idaho)

These details are from [Amazon](https://www.amazon.com/Phylogenetic-Comparative-Methods-Learning-Trees/dp/171958446X).

At the end of Line 26 in the PDF proof: can you replace "constrains" with "constraints".

thanks!
