+++ manuscript/paper.bib
@@ -13,6 +13,7 @@
     number={1},
     pages={1--10},
     year={2018},
+    doi={10.1038/sdata.2018.205},
     publisher={Nature Publishing Group}
 }
 
@@ -21,6 +22,7 @@
   author={Bailey, Norman T.J.},
   year={1964},
   publisher={John Wiley \& Soins},
+  doi={10.2307/2333730},
   address={New York, London, Sydney}
 }
 
@@ -41,10 +43,11 @@
 
 @book{Harmon:2019,
   author={Harmon, Luke J.},
-  title={Phylogenetic Comparative Methods},
-  year={2019},
+  title={Phylogenetic Comparative Methods: Learning from Trees },
+  year={2018},
   eprint={https://lukejharmon.github.io/pcm/pdf/phylogeneticComparativeMethods.pdf},
-  publisher={University of Idaho}
+  isbn = {978-1719584463},
+  publisher={CreateSpace Independent Publishing}
 }
 
 @article{Stadler:2011,
@@ -95,6 +98,7 @@
   number={6097},
   pages={957--960},
   year={2012},
+  doi={10.1126/science.1219669},
   publisher={American Association for the Advancement of Science}
 }
 
@@ -105,6 +109,7 @@
   volume={4},
   pages={181--207},
   year={2016},
+  doi={10.16993/baj.h},
   publisher={IV}
 }
 
@@ -116,6 +121,7 @@
     journal = {Bioinformatics},
     year = {2014},
     volume = {30},
+    doi={10.1093/bioinformatics/btu181},
     pages = {2216-2218},
   }
   
@@ -126,6 +132,7 @@
     journal = {Bioinformatics},
     year = {2018},
     volume = {35},
+    doi={10.1093/bioinformatics/bty633},
     pages = {526-528},
   }
 
@@ -136,6 +143,20 @@
   volume={12},
   number={8},
   year={2017},
+  doi={10.1371/journal.pone.0180908},
   publisher={Public Library of Science}
 }
 
+@article{beast2,
+    doi = {10.1371/journal.pcbi.1006650},
+    author = {Bouckaert, Remco AND Vaughan, Timothy G. AND Barido-Sottani, Joëlle AND Duchêne, Sebastián AND Fourment, Mathieu AND Gavryushkina, Alexandra AND Heled, Joseph AND Jones, Graham AND Kühnert, Denise AND De Maio, Nicola AND Matschiner, Michael AND Mendes, Fábio K. AND Müller, Nicola F. AND Ogilvie, Huw A. AND du Plessis, Louis AND Popinga, Alex AND Rambaut, Andrew AND Rasmussen, David AND Siveroni, Igor AND Suchard, Marc A. AND Wu, Chieh-Hsi AND Xie, Dong AND Zhang, Chi AND Stadler, Tanja AND Drummond, Alexei J.},
+    journal = {PLOS Computational Biology},
+    publisher = {Public Library of Science},
+    title = {BEAST 2.5: An advanced software platform for Bayesian evolutionary analysis},
+    year = {2019},
+    month = {04},
+    volume = {15},
+    url = {https://doi.org/10.1371/journal.pcbi.1006650},
+    pages = {1-28},
+    number = {4},
+}
