### 2021-10-04 a4e39ca
Use lowercase *and* in biblio
