+++ manuscript/paper.bib
@@ -149,7 +149,7 @@
 
 @article{beast2,
     doi = {10.1371/journal.pcbi.1006650},
-    author = {Bouckaert, Remco AND Vaughan, Timothy G. AND Barido-Sottani, Joëlle AND Duchêne, Sebastián AND Fourment, Mathieu AND Gavryushkina, Alexandra AND Heled, Joseph AND Jones, Graham AND Kühnert, Denise AND De Maio, Nicola AND Matschiner, Michael AND Mendes, Fábio K. AND Müller, Nicola F. AND Ogilvie, Huw A. AND du Plessis, Louis AND Popinga, Alex AND Rambaut, Andrew AND Rasmussen, David AND Siveroni, Igor AND Suchard, Marc A. AND Wu, Chieh-Hsi AND Xie, Dong AND Zhang, Chi AND Stadler, Tanja AND Drummond, Alexei J.},
+    author = {Bouckaert, Remco and Vaughan, Timothy G. and Barido-Sottani, Joëlle and Duchêne, Sebastián and Fourment, Mathieu and Gavryushkina, Alexandra and Heled, Joseph and Jones, Graham and Kühnert, Denise and De Maio, Nicola and Matschiner, Michael and Mendes, Fábio K. and Müller, Nicola F. and Ogilvie, Huw A. and du Plessis, Louis and Popinga, Alex and Rambaut, Andrew and Rasmussen, David and Siveroni, Igor and Suchard, Marc A. and Wu, Chieh-Hsi and Xie, Dong and Zhang, Chi and Stadler, Tanja and Drummond, Alexei J.},
     journal = {PLOS Computational Biology},
     publisher = {Public Library of Science},
     title = {BEAST 2.5: An advanced software platform for Bayesian evolutionary analysis},
