### 2021-10-04 164c61a
Remove spurious newlines in code block
