+++ manuscript/paper.bib
@@ -47,6 +47,7 @@
   year={2018},
   eprint={https://lukejharmon.github.io/pcm/pdf/phylogeneticComparativeMethods.pdf},
   isbn = {978-1719584463},
+  doi={10.32942/osf.io/e3xnr},
   publisher={CreateSpace Independent Publishing}
 }
 
