### 2021-10-06 #3173 @majensen
Ok @tresoldi, I'm happy with this. We are nearly there- can you please use Zenodo, FigShare or other service to create an open archive of the repository? There is a short tutorial using Zenodo [here](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729). JOSS prefers that the name of the archive = the name of the paper, and that the archive author list = paper author list.

The service should give you a DOI for the archive - can you post that here when done? Then I will invoke the ancient spell summoning the Associate Editor in Chief.
