+++ setup.py
@@ -49,7 +49,6 @@ setup(
     test_suite="tests",
     tests_require=[],
     url="https://github.com/tresoldi/ngesh",
-    version="0.5",  # remember to sync with __init__.py
+    version="1.0",  # remember to sync with __init__.py
     zip_safe=False,
 )
-
