+++ src/ngesh/__init__.py
@@ -3,7 +3,7 @@ ngesh __init__.py
 """
 
 # Version of the ngesh package
-__version__ = "0.5"
+__version__ = "1.0"
 __author__ = "Tiago Tresoldi"
 __email__ = "tiago.tresoldi@lingfil.uu.se"
 
