+++ manuscript/paper.bib
@@ -21,7 +21,7 @@
   title={The elements of stochastic processes with applications to the natural sciences},
   author={Bailey, Norman T.J.},
   year={1964},
-  publisher={John Wiley \& Soins},
+  publisher={John Wiley \& Sons},
   doi={10.2307/2333730},
   address={New York, London, Sydney}
 }
